Avoiding collisions if intercept directory removed by TextTest, don't use the host name unnecessarily, just use localhost

+++ capturemock/__init__.py
@@ -158,7 +158,8 @@ def commandline():
                 recordEditDir=options.record_file_edits, replayEditDir=options.replay_file_edits,
                 rcFiles=rcFiles, interceptDir=interceptDir)
     subprocess.call(args)
-    shutil.rmtree(interceptDir)
+    if os.path.exists(interceptDir):
+        shutil.rmtree(interceptDir)
     terminate()
